move context out to it own file
