+++ src/content/app/tools/blast/views/blast-form/BlastFormContextContainer.tsx
@@ -1,0 +1,62 @@
+/**
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React, { ReactNode, useState } from 'react';
+
+import useBlastForm from 'src/content/app/tools/blast/hooks/useBlastForm';
+
+type SequenceValidity = {
+  updateSequenceValidity: (index: number | null, status: boolean) => void;
+  removeSequenceValidity: (index: number) => void;
+  sequencesValidity: boolean[];
+};
+
+export const BlastFormContext = React.createContext<
+  SequenceValidity | undefined
+>(undefined);
+
+type Props = {
+  children: ReactNode;
+};
+
+export const BlastFormContextContainer = (props: Props) => {
+  const [sequencesValidity, setSequencesValidity] = useState<boolean[]>([]);
+  const { sequences } = useBlastForm();
+
+  const updateSequenceValidity = (index: number | null, status: boolean) => {
+    const array = [...sequencesValidity];
+    index = index === null ? sequences.length : index;
+    array[index] = status;
+    setSequencesValidity(array);
+  };
+
+  const removeSequenceValidity = (index: number) => {
+    setSequencesValidity(sequencesValidity.filter((value, i) => i !== index));
+  };
+  return (
+    <BlastFormContext.Provider
+      value={{
+        updateSequenceValidity,
+        removeSequenceValidity,
+        sequencesValidity
+      }}
+    >
+      {props.children}
+    </BlastFormContext.Provider>
+  );
+};
+
+export default BlastFormContextContainer;
